+++ frontend/src/components/ErrorSplash.jsx
@@ -1,4 +1,3 @@
-
 import React from 'react'
 import Button from '@mui/material/Button'
 import Typography from '@mui/material/Typography'
@@ -6,11 +5,7 @@ import Box from '@mui/material/Box'
 import koFirefly from '../assets/cuca-marejada.svg'
 import { useNavigate } from 'react-router-dom'
 
-export default function ErrorSplash({
-  message,
-  backlink,
-  backtext,
-}) {
+export default function ErrorSplash({ title, message, backlink, backtext, backaction }) {
   const navigate = useNavigate()
   return (
     <Box
@@ -21,7 +16,7 @@ export default function ErrorSplash({
         p: 2,
       }}
     >
-      <Typography variant="h4">{message}</Typography>
+      <Typography variant="h4">{title || message}</Typography>
       <img
         src={koFirefly}
         style={{
@@ -29,8 +24,13 @@ export default function ErrorSplash({
           marginBlock: '2rem',
         }}
       />
+      {title && message && (
+        <Typography sx={{ m: 3, maxWidth: '30rem', textAlign: 'center' }} variant="body1">
+          {message}
+        </Typography>
+      )}
       <Box sx={{ display: 'flex', gap: 3 }}>
-        <Button variant="contained" color="primary" onClick={() => navigate(backlink)}>
+        <Button variant="contained" color="primary" onClick={backaction || (() => navigate(backlink))}>
           {backtext}
         </Button>
       </Box>