+++ frontend/src/components/LoginDialog.jsx
@@ -26,18 +26,17 @@ function LoginDialog(params) {
     beLoading(true)
     ov.localLogin(username, password)
       .then((response) => {
-        if (!response) {
+        if (response.code === 'VALIDATION_ERROR') {
+          setError(t('LOGIN.VALIDATION_ERROR'))
+          beLoading(false)
+          return
+        }
+        if (response.error){
           setError(t('LOGIN.UNABLE_TO_LOGIN'))
           beLoading(false)
           return
         }
-        console.log(response)
         closeDialog()
-      })
-      .catch((error) => {
-        console.error({ error })
-        setError(t('LOGIN.VALIDATION_ERROR'))
-        beLoading(false)
       })
   }
   return (