+++ CHANGES.md
@@ -2,14 +2,31 @@
 
 ## Unreleased
 
-- UI check API version to match and do a full reload after a countdown
+- New pages: Installation list, Installation details
+- New reusable components:
+  - TableEditor: to show list of items with actions
+  - SimpleTable: to show data sheets
+  - SnackbarMessages: listen messages and show temporary error messages 
+  - ErrorSpash: to show static errors
 - RGPD check required to access
+- UI check API version to match and do a full reload after a countdown
+- Error handling:
+  - Handling ERP errors in API
+  - Sentry integration
+  - UI: ovapi unified error handling
+- Component Test Page visible in development
+- Moved config docs into examples for .env files
 - Improved model validation for vat and email fields
 - Fix: Profile: Retrieve proxy person information for legal entities
 - Profile: Hide empty proxy fields for physical persons
 - Profile: Hide roles (if not `VITE_ENABLE_VIEW_ROLE_IN_PROFILE`)
 - Upgrade Notes:
-    - Added frontend/.env  `VITE_ENABLE_VIEW_ROLE_IN_PROFILE`
+    - Added frontend/.env variables:
+        - `VITE_ENABLE_VIEW_ROLE_IN_PROFILE`
+        - `VITE_ENABLE_BREAKPOINT_INDICATOR`
+    - Added .env variables:
+        - `SENTRY_DSN` (by default no sentry loggin)
+        - `ENV` (by default 'development')
 
 ## 0.1.0 (2023-11-03)